kern: forbid(unsafe_op_in_unsafe_fn).
This commit is like 90% unsafe keywords and "Safety:" comments, and 10%
something apparently unrelated. Namely, I've changed the
`with_task_table` routine to use an atomic bool to detect reentrant use,
and changed it to "safe." This is not as unrelated as it may appear:
enabling `unsafe_op_in_unsafe_fn` caused a bunch of cases where code
looked like this...

    unsafe {
        with_task_table(|tasks| { stuuuuuff });
    }

Because Rust infers an `unsafe` block around the contents of any
closures declared in an `unsafe` block, that allowed arbitrary unsafe
operations in the body of that closure -- which in practice can be a
screen or two in length. Making `with_task_table` safe eliminates the
outer unsafe block and dodges this footgun, which I still consider to be
a design flaw in the language. Oh well.

Anyway, to make `with_task_table` safe required the introduction of
atomics into the kernel, which in turn required some gymnastics to
support Cortex-M0, where some atomic operations are silently missing
from core but without any cfg help from rustc in detecting this,
speaking of design flaws in the language.