kern: portable with_task_table
This moves with_task_table out of arch, now that we can locate the task
table in a portable and predictable way (because there's a static symbol
for it).

By relying on the static symbol at link time instead of reading
variables, this reduces kernel size by a bit (around 100 bytes).