kern: fix subtle long-standing asm bug
This is a good one. The assembly sequences that do context save on the
way into the kernel are expected to leave a task pointer in a
predictable location (r1) for use by the Rust code they wind up calling.
They've been doing it wrong.

On both ARMv6-M and ARMv7/8-M, we've been passing an _interior_ pointer
to the task, pointing just after the register save area, as a result of
using store-multiple-increment-ascending to do the context save.

Hilariously, the Rust code has been _hiding this fact_ because of the
way it uses the task pointer: it turns it into a task _index_ using a
divide, which effectively rounds it down!

One of my in-flight changes stops using the divide, which revealed this
bug.