thermal: use Jefe state, not sequencer state
This change causes the thermal loop to respect the system state
available from Jefe (and produced by the sequencer) rather than directly
hassling the sequencer, breaking the dependency between the two tasks.

That's mostly a mechanical change, but there is one subtlety: the
thermal task has been implicitly relying on the sequencer _not
responding to IPCs_ before it reaches A2. Basically, the thermal task
was sending an IPC to the sequencer (indirectly, via the BSP), which
would cause it to block until the sequencer entered its IPC handling
loop in A2.

Jefe, on the other hand, is always live, so you can ask Jefe about the
system state _well before_ A2 and get a response (0). This would have
caused the thermal task to go talk to the fan controller etc. and become
unhappy.

I've rearranged things to make the relationship between power state and
starting the control loop explicit, rather than implicit in which BSP
operations do or do not block.