Apply serde(deny_unknown_fields) to all config structs
serde+toml by default will tolerate arbitrary unknown fields, producing
ground fertile for errors such as typos or pasting in the wrong section.
While fixing serde+toml to not do this by default is _awfully tempting_
this change instead goes through and decorates all of our config structs
with annotations to suppress the behavior.

While I was at it, I have applied rename_all = kebab-case to all config
structs that (1) were missing it and (2) don't have underscores in any
field names, so it's a no-op. This is to prepare for the addition of
other field names in the future. I'll be going through and kebabing the
rest in a bit.