Support for Meanwell application (#620)
This adds support for a Meanwell application definition for Gimletlet,
which allows for a Gimletlet to be easily used to manage Meanwell
power supplies on a benchtop. There is nothing particularly
Meanwell-specific (or indeed even power-specific) here; this is really
just a thin veneer on the GPIOs present on the J16 header on the
Gimletlet, allowing them to be used to manage external power supplies
without forcing the user to wade into the intricacies of GPIO pin
configuration.  Note that the Meanwell application also indicates a
different LED pattern (a quick blue LED blink, once per second) to
make it clear from across a room that a particularly Gimletlet is
serving to manage Meanwell supplies.