lib-dice: Add new library with types for DICE artifacts and certs.
The current cert implementation is limited to the creation of a self
signed cert for the DeviceId & the Alias cert. Cert creation is as
simple as possible with a template for each cert (DER) that is populated
with fixed length data like the platform serial number, certificate
serial number, public key & signature. Future work will address
programatic generation of these templates at build time to replace the
currently static ones.