lpc55/chip.toml: Add memory regions for DICE artifact handoff.
The memory used in this commit is SRAM intended for use by USB. The
choice of using this memory region was mostly driven by convenience and
pivoting to a different range should be straight forward. The size of
this region was chosen based on the estimated size of the data passed.
The DeviceId and Alias cert are each ~600 bytes and the seed required to
reconstruct the Alias key is 32 bytes. All of this fits comfortably in
2k.