vsc*: adjust inline/monomorph decisions
The general theme here is:

- If a generic function is going to be instantiated in a lot of
  different forms -- in the vsc PAC case, once for every unique register
  -- try to inline it.
- If a function is going to be inlined, try to make sure its body is
  relatively short, and probably implemented in terms of non-inlined
  parts.
- Read the compiler output to check your inline(always)/inline(never)
  assumptions, because sometimes inlining is the wrong thing to do.

So, in this case, we had read/write routines that were generic on the
register they were accessing. This provides useful type safety in the
API, but was generating an awful lot of code. Here they've been
rewritten into a non-generic core and thin generic wrapper, which still
provides the type safety to API consumers, but reduces the size of
generated code and makes the PhyRw trait much simpler to implement.

This knocks 49 kiB off monorail and about 1 kiB off net.