sidecar: hack to allow larger bitstreams for now
This allows sidecar to use both banks of the STM32H7 flash for its
images, which in turn allows the use of FPGA bitstreams that are larger
than ~100kiB. I've tested this with up to 512 kiB of bitstreams and
there's space to spare.

Note that this destroys sidecar's ability to do online updates or have
redundant firmware images. So, this is explicitly intended as a
short-term hack to unblock bringup while we get the external flash
online.