Enable SP VLAN on remaining products (#703)
The SP VLAN feature (described in RFD 250 and initially implemented in #510)
lets the SP pretend to be two independent MACs.  It was previously only enabled
for Gimlet, but we want it turned on for all Oxide products.

I also consolidated features and `stacksize`, setting the latter to the max
value that wouldn't bump RAM to 32768.

Stack margin looks reasonable: 768 on Gimlet, 744 on PSC, 896 on Sidecar.