stm32h7: allow optional mapping into AXI SRAM
By default, apps on stm32h7 are mapped into a 128 kiB RAM on the
processor's local bus. This is great because it's zero-wait-state.

However, there are a lot more RAMs than that. This adds an alternate
memory file for mapping all of an application into the 512 kiB AXI SRAM,
which has more wait states (though not many more) but is 4x larger.

This is in lieu of a smarter memory allocator that can distribute across
RAMs based on their properties, which we'll want to do eventually.