userlib: add cortex-m dep to force inline-asm feature.
We were rebuilding PACs twice for most applications, once for tasks
where the inline-asm feature is set somewhere in their dependency graph,
and once for tasks where it isn't. The inline-asm feature is thankfully
going away in cortex-m 0.8 (it should be a build-time capability
detection rather than a user-visible feature), but for now, we'll just
force it on in the one library that all tasks include: userlib.

This knocks a minute or more off a clean build for stm32h7, where the
PAC is particularly expensive.