Fill gaps in binaries with `0xff` (#724)
By default, objcopy will fill gaps in sections with `0x00`. The flash
programming sequence on the STM32 will not program the gaps and will
leave them in the default state of `0xff`. This results in a mismatch
between the generated binary and flash. From a running program
perspective this is fine but it results in incorrect calculations of
expected values in flash. Fix this by filling the gap with `0xff`.

Co-authored-by: Laura Abbott <laura@oxide.computer>