mgmt-gateway: Update to new MGS serialization of large messages
Update payloads and serial console packets are now serialized as raw
data _after_ a hubpack-encoded message, instead of the data being
embedded as part of the hubpack message. Formerly we were required to
allocate several moderately-sized arrays on the stack and copy data
unnecessarily; now the raw data is borrowed from the receive buffer we
pass to `net`, which requires no extra copies or stack allocation of
large arrays.