lib-dice: Implement builder pattern for certs.
This moves all of "setters" from the Cert trait into a new CertBuilder
trait. Implementations of the builder trait are added for the two
certs that we create. Builders are created by their respective 'new'
functions that take all required fields as parameters. Certs are then
instantiated by calling the 'sign' method on the builder.