simplify duplicate job removal
We would get duplicate builds for PRs because we want to check on
master, but also on pull requests. Due to using 'push' for master, we'd
end up with these duplicate jobs, and used this action to sort out the
duplicates.

This change simplifies this by only doing the 'push' actions on master,
and not on PR branches.