Dynamically determine the clock speed for the LPC55
The default clock speed for the LPC55 is 48MHz. We eventually want
to run at 96MHz. The easiest way for our system to do that is via
the CMPA setting. We still need to be able to run on 48MHz systems
as well. Work around this by dynamically determining the clock
speed at runtime instead of relying on a compile time constant.