Gimlet Rev C + MAX5970 sensor support (#841)
This work includes both support for Gimlet Rev C and the voltage and
current sensing on the MAX5970.  This latter change necessitated a
slight change to the device metadata, as the MAX5970 is not a PMBus
device. As a result, to see PMBus devices via Humility's
PMBus-specific functionality (e.g., "humility pmbus") after this
change, you will need Humility version 0.9.0 or later.