net: add RX activity watchdog.
We're hitting an intermittent bug on the stm32h7 where the Ethernet MAC
stops delivering packets from the MTL, without any status bits to
indicate it or any obvious way to get it unstuck. Resetting the device
works, however!

This changes the netstack to reset itself (by panicking) if no traffic
has emerged from the MAC within a time window (currently 60 seconds).
Note that this should cover any traffic on the wire, _not just_ traffic
addressed to the node. The watchdog interval needs to be short enough to
make it practical for an operator to get control over a wedged machine,
but doesn't have to be "machine-fast" because we don't expect this to
happen very often. (If it turns out to happen a lot, we should
reconsider this.)

It would be nice to reset just the Ethernet driver instead of losing
netstack state, which includes any packets received to sockets but not
yet handed out to tasks. Maybe later.