net: simplify server_vlan setup
Previously this was following a large number of parallel arrays and
zipping them; because zipping large numbers of iterators is annoying
using only core, it put them in mutable locals and stepped them
manually. This has some drawbacks:

- It generated a lot of mutable locals to keep track of
- It allowed for iterators _not being stepped_ on iterations of the
  loops, knocking the zip out of step due to mistakes

This change reduces the number of parallel arrays of things that we zip
to 3, which makes it practical to just zip them, removing the potential
for errors.

I've also switched the gradual array initialization of the state array
to use heapless::Vec, which makes the code simpler (IMO) and should
reduce copies during init by removing the need to unwrap an array of
Option.