net: eliminate NetServer, collapse server implementations.
Previously, while we had managed to eliminate a lot of the logic
duplication between the vlan and basic servers, there was still a
nontrivial amount of duplicate server logic -- particularly, logic for
things like LargePayloadBehavior, handling the Exhausted condition, etc.
This logic was in the impls of the NetServer trait.

This commit eliminates the NetServer trait and generalizes the vlan impl
of it to also work for the basic, non-VLAN case, by specialization. The
vlan and basic server modules still exist, but are only responsible for
providing the (now extended) DeviceExt impl for their ethernet driver
wrappers, and a collection of hooks to customize receive logic.