Check for out-of-sync incoming requests
Previously, a misbehaving host could cause us to become deadlocked with
hardware flow control enabled: we did not attempt to read from the RX
FIFO if we had data to send, so a host that tried to send to
back-to-back requests without receiving the first response could end up
with both ends trying to transmit and neither able to make progress.

Now, if we have filled the TX FIFO, we'll check for an incoming packet.
If we receive one, we'll discard our old (presumably partially-sent)
response and handle the new request.