abi: remove TaskDesc/RegionDesc and friends
These were in abi because, originally, the kernel was totally separately
compiled from tasks and interpreted a series of these structs (at a
predictable address) at startup to create tasks. This prevented a bunch
of optimizations and made kernel RAM usage hard to predict. That
behavior stopped about a year ago.

They've been in abi since then because they were being used to
communicate between xtask dist and kern/build.rs in serialized form.
That stopped yesterday.

Moving these into the kernel will make it a lot easier to iterate on
them, and in particular lifting the zerocopy/serde restriction allows
the use of reference and pointer types.