Dice mfg cmds (#833)
* dice: reclaim memory by handing off common data through a shared memory region

Sharing memory between tasks like this requires we map it read-only
which AFAIK isn't currently possible. This will need to be added to
hubris.

* stage0: Introduce scope changes in the DICE artifact generation to save ram.

This seems to be the most reliable method for getting the compiler to
reuse stack space. The end result isn't how I'd choose to structure the
code but I need this memory back.

* lib-dice: Add mfg interface, move common types to mfg lib.

New mfg lib provides types for blobs like the CSR and Certs where one end
of the mfg communication channel can't know the appropriate size in
advance. Specifically the mfg side won't know how big the CSR is, and
the RoT won't know the size of the DeviceId or intermediate certs. The
platform serial number is part of the API for the mfg commands so that's
been moved into this library as well.

* dice: Add CSR generation & mfg cmds, reorg dice crate feature.

Adds a features for two operational modes:
- dice-self - generate self signed deviceid cert
- dice-mfg - use usart0 to communicate with external tools & CA to get a
  deviceid cert signed by a CA

* dice: Write data received from manufacturing commands to flash.

Pull data from flash on boot & populate handoff memory.

* lib-dice: fix serial number length

* dice: drop OU from subject / issuer, tighten up constraints & extensions