Support NVME BMC devices with bad mux (#888)
Due to https://github.com/oxidecomputer/hardware-gimlet/issues/1804, we can't
talk to M.2s when they are unpowered.  We work around this in the `thermal` task
by asking the `MAX5970` for its status; however, `humility validate` doesn't
have any support for that kind of special-casing.

This PR adds a new `m2_hp_only`, which is ignored during `humility validate`.

In addition, it renames `nvmebmc` -> `nvme_bmc` so that the existing driver is
used for `humility validate` (on good devices).