Fix salty dep by adding branch.
Previously salty used an explicit git rev in the Cargo.toml. This isn't
great practice since it prevents easy updating and allows you to
reference a commit that isn't on a branch, and is thus subject to being
arbitrarily GC'd by Github at any time, breaking the build.

The conversion to workspace dependencies dropped the rev with the
rationale that it appeared in the Cargo.lock. This was true, but, the
rev was also not on salty's default branch. Cargo only checks out the
branch you request. So, on machines that hadn't previously built the
explicit-rev version and still had a salty checkout cached, the build
now fails because the locked commit is absent.

This switches the dependency to name the branch in our fork where our
desired commits live. This ensures that we can cargo update to anything
new we push on that branch, and that the commits actually get checked
out by Cargo.

I've audited the other packages that were using rev= dependencies before
the workspace change, and in all cases, their commits were on their
default branches. So only salty needs to change here.