[sprot] Try to make error handling more idiomatic
Overall, the behavior of the protocol should not have changed!

The overriding goal of this commit is to simplify error handling while
making it more idiomatic. The primary methodology used was "parse, don't
validate", meaning encoding knowledge about what was already discovered
in the type system, so that we don't have to redo checks about length,
or reparse the `MsgHeader`. The mechanisms to implement this methodology
are the two types: `RxMsg` and `TxMsg`, and the corresponding types
`VerifiedRxMsg` and `VerifiedTxMsg`, which ensure that a complete
message with a valid `MsgHeader` has been read into `rx_buf` or a valid
message  has been serialized into `tx_buf`, respectively.

The above changes rippled through the rest of the code, and allowed
removing some extra checks around message/buffer sizes. There were a
handful of other changes including:

 * Some function signatures changed to use the new types
 * `MsgError` was renamed to `SprotError`
 * `UpdateError` variants are now mapped 1:1 into `SprotError` variants.
While this adds a bunch of extra variants, it gives us more information
about what went wrong for debugging, and also allows us to expose a
single error type from Sprot.
 * Some extra ringbuf trace messages for debugging were added.
 * Some standalone functions related to buffer access were superceded
   by their `TxMsg`/`RxMsg` methods.
 * `ErrorRsp` messages now always include an underlying 1-byte code