Add separate monitoring for PHY link flapping (#970)
In some cases, the VSC7448 link is going to a PHY, so link status is less
meaningful.  This is mostly relevant for Sidecar ports 40-43, which use a
VSC8504 as a QSGMII → 4x SGMII converter before going off-board; the link
between the VSC7448 and VSC8504 should _always be up_, because they're on the
same board.

This PR adds a separate `phy_link_down_sticky: bool` to the `struct
PortCounters`, which returns the **PHY's** opinion on whether the link has gone
done.  We want this to be stateless (i.e. explicitly cleared by
`reset_port_counters`), but the PHY registers are self-clearing; to work around
this, we add a separate cache in the Monorail `ServerImpl`.

(this cache will be lost if the `monorail` task restarts, but it would
reconfigure the PHYs anyways in that case, so that seems fine)