Add support for a `status` message to update-api
Currently the status only reports on the currently active slot. The
value is correct for the lpc55, but there is a placeholder for the
stm32h7. Future commits will provide more detail in the `UpdateStatus`
value returned such as which version is in which slot, the image ids,
and image hashes. We expect to remove the `current_version` API, such
that it is subsumed as part of `status`.

The API has been plumbed through the update servers and sprot servers on
both RoT and SP.