Add transceivers to thermal control loop (#1010)
This PR adds the 32x Sidecar transceivers to the thermal loop managed by
`task-thermal`.

This is non-trivial: until now, the thermal loop assumed that all of its sensors
would be known I2C devices, and that it would be responsible for polling them
itself.

Transceivers break these assumptions:
- A given transceiver isn't fixed; users may insert hardware with different
  specifications and thermal limits
- Transceivers, while _technically_ I2C devices, aren't read through our normal
  I2C code; they're read through a pair of FPGAs on the front IO board.

The fix is to add a new category of sensors to the thermal loop.  "Dynamic
inputs" represent sensors which are both
- Configured at runtime with a particular thermal model, and
- Fed data by a **different task**

The different task is our `transceivers` task, which is the liaison to the
FPGAs.  When a transceiver is detected, this task picks a thermal model and
interface*, then periodically polls and sends temperature data to `thermal`.

There are also a few other drive-by changes:

- Switching to strong types for `Physical` vs `Logical` transceiver ports
- Polling the _actual_ power state of Sidecar, instead of the sequencer
  configuration
- **Enabling the thermal loop for Sidecar**, with the PID tuning from Gimlet (as
  a temporary value)

Finally, the `sensors` task and build system is made slightly more generic: it
now accepts a set of extra sensors (other than I2C devices) in the TOML file
(under `config.sensor.devices`), and does codegen for them.

--------------------------------------------------------------------------------

*Right now, the interface is selected based on byte 0 of the memory map, which
contains an identifier code; the thermal model is picked arbitrarily and
somewhat conservatively.  Eventually, we expect host software to send us this
data before enabling a transceiver.