Add timestamps to sensor readings (#1023)
This PR adds timestamps to readings within the `sensors` task, which must be
supplied when calling `Sensor.post` and `Sensor.nodata`

In addition, sensor data is reorganized into a struct-of-arrays which stores the
most recent reading **and** the most recent error, along with a marker
indicating which (of the two) is most recent.  Using a struct-of-arrays saves a
significant amount of RAM (which would otherwise be lost to padding between
`struct` members).

The existing `Sensor.get` API is unchanged to preserve compatibility with
Humility; the PR adds `Sensor.get_reading` which returns both the value and the
timestamp.