kern: record failure message in predictable place.
This should make kernel crashes and panics easier to inspect using a
debugger, by recording state in predictably-named statics. This replaces
the formerly-configurable panic handler, which in practice was always
set to itm, and which in practice was rarely useful because of
inconsistent ITM support. This approach has the distinct advantage of
working on systems without ITM, like M0s.