dice: Insert a PUF derived key under the first DICE key.
This adds a new certificate template for the self signed configuration
and re-purposes the deviceid CSR template used for the manufacturing
process for the new credential.

The approach taken trades off additional memory usage (we've added
another key & cert) in exchange for less complexity in the logic and
less code duplicated across the stage0 mfg code. The additional memory
usage comes from flattening out the function call hierarchy and
returning the certs from the manufacturing process to the top level
`dice::run()` function.