Sprot Enhancements
These changes started as a continuation of
https://github.com/oxidecomputer/hubris/pull/965. In particular,
the buffer wrappers `TxMsg` and `RxMsg` in the sprot API owned
the underlying buffers. However, the `Verified` versions of these types
that got created when a message was serialized  into a tx buffer
or parsed successfully from an rx buffer did not take ownership,
but also did not mutably borrow from the owned types. This meant that
the underlying buffers could be manipulated while the Verified versions
existed, which somewhat negates the benefit of these types. This
PR rectifies this by making both versions borrow mutably from an
underlying buffer outside the types. However, on the SP side,
due to retry loops we sometimes have to not directly convert in a
typestate manner. On the RoT however, the types work  as expected.

This PR grew significantly, as it was being used to test RoT upgrade
from faux-mgs via control-plane-agent. It was noticed that on both the
master branch and the branch with the buffer changes (only on the RoT
side at first), that update would repeatedly fail due to overruns. While
debugging, I noticed invalid headers showing up at the RoT side while
querying status during update. This is something that faux-mgs does
interspersed with block writes. I theorized that this resulted from  the
SP side timing out and sending a new request to the RoT while the RoT
was sending back a response from the prior message. As an experiment,
to facilitate my understanding of the code, and enable better debugging
via ringbuf, I broke the existing `pio` loop down into a state machine
that separated out the reading the request part of the code, where
the RoT would always clock out zero bytes, from the responding part of
the code, where the RoT would clock out the response after asserting
`ROT_IRQ`. This allowed both the RoT and the SP to detect when they were
desynchronized, and also helped ensure that a response to the SP
would always be for the latest request without the need for sequence
numbers at the sprot api level.

There were a number of other beneficial changes on the RoT side that
fell out of this restructuring into a state machine:

* The RoT can detect when its desynchronized in a straightforward manner
as it will receive non-zero bytes while clocking out a response.
* The RoT side of sprot can now log it's state transitions, when it sees
a CSn pulse, and when it thinks it's desynchronized
* The stats reported by the RoT are also enhanced, as the RoT has more
information about the state of the protocol and what the SP should
be doing because we constrain it to request/reply one message at a time.
* We only have to wait for CSn asserted interrupts, as we know when
to go directly into a polling code.
* The handler code is further decoupled and simplified, such that
 "no message" detected requests don't call into the handler,
and neither do overrun or underruns, as stats are now handled
directly in the IO code.

On the SP side there were also a number of improvements outside
the RxMsg/TxMsg related updates:
 * RAII is now used with the spi API, such that manual release
of the lock is no longer needed.
 * Request sending is simplified and wrapped in a function
 * IRQ related code was broken out into functions and comments moved to
function headers
 * Duplicate checks for error conditions in `do_send_recv_retries`
were removed from `rot_sink`, simplifying it.

There was also an extra constraint put on the rot_sink test: It now only
supports buffers of two bytes or larger. This was just for simplicity,
but it can be reverted.

It should be noted that in one point of view, the changes to explicitly
specify reading and writing halfs of the RoT server can be seen as
adding complexity. The RoT no longer just handles requests and responds
to them, it now has to manage the current state of the RoT. However, the
current, de-facto implementation on master is that:
 1. The Sp sends a request by asserting CSn and De-asserting CSn when it
is done
 2. The RoT handles that request
 3. The RoT asserts `ROT_IRQ`, which triggers the SP to assert CSn and
 clock zero bytes until it receives a full response to its request.
 4. The RoT de-asserts `ROT_IRQ` when it sees CSn de-asserted.

The state machine as written explicitly encodes this behavior rather
than leaving it implicit in a single `pio` loop. Any flow outside
this happy path, can result in the RoT returning a response to a prior
message, or the SP triggering a CSn pulse due to it not expecting
`ROT_IRQ` to be asserted while it is sending a request. By having the
RoT understand this flow, it can deal with each corner case in turn.
Since other bugs and corner cases may arise, it also seems useful for
debugging to be able to detect on both the RoT and the SP that we are
outside the happy path.