add in situ dump support (#1055)
This work adds support for in situ dumping, whereby a dump can be
written to auxiliary memory within Hubris (currently sram2, sram3, and
sram4 on Gimlet).  The memory to dump is recorded within Hubris; an
outside entity reads the memory to dump, compresses it, and write it
to a designated dump area.  Given that Gimlet's SP is SWD-controlled
by the RoT, this allows for dumps of the SP to be initiated by the
network, taken by the RoT -- and later retrieved over the netword.
("humility dump" has been extended to handle this process
automatically.)