xtask: calculate non-overlapping SAU ranges
When USB SRAM was added as an additional allocatable section,
the SAU range algorithm simply copied that range into an
additional SAU range.  Unfortunately, there is an implicit SAU
range added for all of the entire non-secure peripheral address
space.

From the ARMv8M spec: "An address that matches multiple SAU regions
is marked as Secure and not Not-secure callable regardless of the
attributes specified by the regions that matched the address."

Instead, use the rangemap crate to build up non-overlapping ranges
from the allocatable regions and implicit ranges.