When starting an LPC55 update, erase block 0, but do not write any data.
Some testing and research is needed, but here is a draft PR.

On the LPC55, stage0 handles the case where block 0 is erase, i.e.
block 0 is not zero or any other pattern.

It's important to invalidate the header as the first step in updating
the bank. However, it's the erase/write cycle that causes wear in most flash parts.
LPC55's data sheet claims 10k erase/write cycles (IIRC).
So, a pathological update fail/retry loop where each iteration
takes 10 seconds could burn out the flash in just over a day.

If the first step is only erase, and the LPC55 will not erase an erased block,
or we can test for that, or erase on top of erase doesn't cause wear,
then not writing any data to block
zero would eliminate wear in this pathological case.